# aspnetcore-server/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "Room Reservation API",
    "description" : "A Simple IP Address API",
    "version" : "0.1.9"
  },
  "servers" : [ {
    "url" : "http://localhost:5000/api"
  } ],
  "paths" : {
    "/meetingRooms" : {
      "get" : {
        "tags" : [ "meeting room" ],
        "summary" : "Returns a list of rooms",
        "responses" : {
          "200" : {
            "description" : "An array of meeting rooms",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/MeetingRoom"
                  },
                  "x-content-type" : "application/json"
                }
              }
            }
          },
          "404" : {
            "description" : "Rooms not found"
          }
        }
      },
      "post" : {
        "tags" : [ "meeting room" ],
        "summary" : "Creates a new room",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MeetingRoom"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MeetingRoom"
                }
              }
            }
          },
          "409" : {
            "description" : "Conflict"
          }
        }
      }
    },
    "/meetingRooms/{meetingRoomId}" : {
      "get" : {
        "tags" : [ "meeting room" ],
        "summary" : "Returns a meeting room by meetingRoomId",
        "parameters" : [ {
          "name" : "meetingRoomId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MeetingRoom"
                }
              }
            }
          },
          "404" : {
            "description" : "Room not found"
          }
        }
      },
      "put" : {
        "tags" : [ "meeting room" ],
        "summary" : "Updates an existing room",
        "parameters" : [ {
          "name" : "meetingRoomId",
          "in" : "path",
          "description" : "ID of a meeting room",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MeetingRoom"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "Updated"
          },
          "400" : {
            "description" : "Bad Request"
          },
          "404" : {
            "description" : "Room not found"
          }
        }
      },
      "delete" : {
        "tags" : [ "meeting room" ],
        "summary" : "Deletes a room",
        "parameters" : [ {
          "name" : "meetingRoomId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Deleted"
          },
          "404" : {
            "description" : "Room not found"
          }
        }
      }
    },
    "/meetingRooms/image/{imageName}" : {
      "get" : {
        "tags" : [ "meeting room" ],
        "summary" : "Returns room by an image name",
        "parameters" : [ {
          "name" : "imageName",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "A room object",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MeetingRoom"
                }
              }
            }
          },
          "404" : {
            "description" : "Room not found"
          }
        }
      }
    },
...
    "/meetingRoomImage" : {
      "post" : {
        "tags" : [ "meeting room image" ],
        "summary" : "Creates a new image",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MeetingRoomImage"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MeetingRoomImage"
                }
              }
            }
          },
          "409" : {
            "description" : "Conflict"
          }
        }
      }
    },
    "/meetingRoomImage/{imageName}" : {
      "get" : {
        "tags" : [ "meeting room image" ],
        "summary" : "Returns meeting room Image by imagename",
        "parameters" : [ {
          "name" : "imageName",
          "in" : "path",
          "description" : "imagename of a meetingRoomImage",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MeetingRoomImage"
                }
              }
            }
          },
          "404" : {
            "description" : "Meeting Room Image not found"
          }
        }
      },
      "put" : {
        "tags" : [ "meeting room image" ],
        "summary" : "Update an existing room",
        "parameters" : [ {
          "name" : "imageName",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MeetingRoomImage"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "Updated"
          },
          "400" : {
            "description" : "Bad Request"
          },
          "404" : {
            "description" : "Room not found"
          }
        }
      },
      "delete" : {
        "tags" : [ "meeting room image" ],
        "summary" : "Deletes a room",
        "parameters" : [ {
          "name" : "imageName",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Deleted"
          },
          "404" : {
            "description" : "RoomImage not found"
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "MeetingRoom" : {
        "type" : "object",
        "properties" : {
          "meetingRoomId" : {
            "type" : "integer",
            "readOnly" : true
          },
          "meetingRoomName" : {
            "type" : "string"
          },
          "meetingRoomEmail" : {
            "type" : "string",
            "format" : "email"
          },
          "meetingRoomLocation" : {
            "$ref" : "#/components/schemas/Location"
          },
          "locationId" : {
            "type" : "integer"
          },
          "meetingRoomImage" : {
            "$ref" : "#/components/schemas/MeetingRoomImage"
          },
          "meetingRoomReservations" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/MeetingRoomReservation"
            }
          }
        },
        "example" : {
          "meetingRoomLocation" : {
            "meetingRooms" : [ null, null ],
            "city" : {
              "cityName" : "cityName",
              "cityId" : 1
            },
            "locationId" : 6,
            "workingSpots" : [ {
              "workingSpotId" : 5,
              "locationId" : 5,
              "location" : null,
              "workingSpotReservations" : [ {
                "workingSpotId" : 7,
                "workingSpot" : null,
                "workingSpotReservationId" : 2,
                "reservationDate" : "2000-01-23",
                "user" : {
                  "jobId" : 2,
                  "password" : "",
                  "reservations" : [ {
                    "meetingRoom" : null,
                    "meetingRoomReservationId" : 4,
                    "meetingRoomId" : 1,
                    "meetingDate" : "2000-01-23",
                    "startTime" : "startTime",
                    "endTime" : "endTime",
                    "meetingTopic" : "meetingTopic",
                    "user" : null,
                    "userId" : 7
                  }, {
                    "meetingRoom" : null,
                    "meetingRoomReservationId" : 4,
                    "meetingRoomId" : 1,
                    "meetingDate" : "2000-01-23",
                    "startTime" : "startTime",
                    "endTime" : "endTime",
                    "meetingTopic" : "meetingTopic",
                    "user" : null,
                    "userId" : 7
                  } ],
                  "surname" : "surname",
                  "name" : "name",
                  "job" : {
                    "jobId" : 3,
                    "jobTitle" : "jobTitle"
                  },
                  "userId" : 9,
                  "email" : ""
                },
                "userId" : 1
              }, {
                "workingSpotId" : 7,
                "workingSpot" : null,
                "workingSpotReservationId" : 2,
                "reservationDate" : "2000-01-23",
                "user" : {
                  "jobId" : 2,
                  "password" : "",
                  "reservations" : [ {
                    "meetingRoom" : null,
                    "meetingRoomReservationId" : 4,
                    "meetingRoomId" : 1,
                    "meetingDate" : "2000-01-23",
                    "startTime" : "startTime",
                    "endTime" : "endTime",
                    "meetingTopic" : "meetingTopic",
                    "user" : null,
                    "userId" : 7
                  }, {
                    "meetingRoom" : null,
                    "meetingRoomReservationId" : 4,
                    "meetingRoomId" : 1,
                    "meetingDate" : "2000-01-23",
                    "startTime" : "startTime",
                    "endTime" : "endTime",
                    "meetingTopic" : "meetingTopic",
                    "user" : null,
                    "userId" : 7
                  } ],
                  "surname" : "surname",
                  "name" : "name",
                  "job" : {
                    "jobId" : 3,
                    "jobTitle" : "jobTitle"
                  },
                  "userId" : 9,
                  "email" : ""
                },
                "userId" : 1
              } ]
            }, {
              "workingSpotId" : 5,
              "locationId" : 5,
              "location" : null,
              "workingSpotReservations" : [ {
                "workingSpotId" : 7,
                "workingSpot" : null,
                "workingSpotReservationId" : 2,
                "reservationDate" : "2000-01-23",
                "user" : {
                  "jobId" : 2,
                  "password" : "",
                  "reservations" : [ {
                    "meetingRoom" : null,
                    "meetingRoomReservationId" : 4,
                    "meetingRoomId" : 1,
                    "meetingDate" : "2000-01-23",
                    "startTime" : "startTime",
                    "endTime" : "endTime",
                    "meetingTopic" : "meetingTopic",
                    "user" : null,
                    "userId" : 7
                  }, {
                    "meetingRoom" : null,
                    "meetingRoomReservationId" : 4,
                    "meetingRoomId" : 1,
                    "meetingDate" : "2000-01-23",
                    "startTime" : "startTime",
                    "endTime" : "endTime",
                    "meetingTopic" : "meetingTopic",
                    "user" : null,
                    "userId" : 7
                  } ],
                  "surname" : "surname",
                  "name" : "name",
                  "job" : {
                    "jobId" : 3,
                    "jobTitle" : "jobTitle"
                  },
                  "userId" : 9,
                  "email" : ""
                },
                "userId" : 1
              }, {
                "workingSpotId" : 7,
                "workingSpot" : null,
                "workingSpotReservationId" : 2,
                "reservationDate" : "2000-01-23",
                "user" : {
                  "jobId" : 2,
                  "password" : "",
                  "reservations" : [ {
                    "meetingRoom" : null,
                    "meetingRoomReservationId" : 4,
                    "meetingRoomId" : 1,
                    "meetingDate" : "2000-01-23",
                    "startTime" : "startTime",
                    "endTime" : "endTime",
                    "meetingTopic" : "meetingTopic",
                    "user" : null,
                    "userId" : 7
                  }, {
                    "meetingRoom" : null,
                    "meetingRoomReservationId" : 4,
                    "meetingRoomId" : 1,
                    "meetingDate" : "2000-01-23",
                    "startTime" : "startTime",
                    "endTime" : "endTime",
                    "meetingTopic" : "meetingTopic",
                    "user" : null,
                    "userId" : 7
                  } ],
                  "surname" : "surname",
                  "name" : "name",
                  "job" : {
                    "jobId" : 3,
                    "jobTitle" : "jobTitle"
                  },
                  "userId" : 9,
                  "email" : ""
                },
                "userId" : 1
              } ]
            } ]
          },
          "meetingRoomId" : 0,
          "meetingRoomName" : "meetingRoomName",
          "locationId" : 1,
          "meetingRoomImage" : {
            "imageWidth" : 7.457744773683766,
            "image" : "",
            "meetingRoom" : null,
            "meetingRoomId" : 1,
            "name" : "name",
            "meetingRoomImageId" : 6
          },
          "meetingRoomReservations" : [ null, null ],
          "meetingRoomEmail" : ""
        }
      },
...
      "MeetingRoomImage" : {
        "type" : "object",
        "properties" : {
          "meetingRoomImageId" : {
            "type" : "integer",
            "readOnly" : true
          },
          "name" : {
            "type" : "string"
          },
          "image" : {
            "type" : "string",
            "format" : "byte"
          },
          "imageWidth" : {
            "type" : "number"
          },
          "meetingRoomId" : {
            "type" : "integer"
          },
          "meetingRoom" : {
            "$ref" : "#/components/schemas/MeetingRoom"
          }
        },
        "example" : {
          "imageWidth" : 7.457744773683766,
          "image" : "",
          "meetingRoom" : null,
          "meetingRoomId" : 1,
          "name" : "name",
          "meetingRoomImageId" : 6
        }
      },
      "City" : {
        "type" : "object",
        "properties" : {
          "cityId" : {
            "type" : "integer",
            "readOnly" : true
          },
          "cityName" : {
            "type" : "string"
          }
        },
        "example" : {
          "cityName" : "cityName",
          "cityId" : 1
        }
      },
...
      "Location" : {
        "type" : "object",
        "properties" : {
          "locationId" : {
            "type" : "integer"
          },
          "city" : {
            "$ref" : "#/components/schemas/City"
          },
          "meetingRooms" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/MeetingRoom"
            }
          },
          "workingSpots" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/WorkingSpot"
            }
          }
        },
        "example" : {
          "meetingRooms" : [ null, null ],
          "city" : {
            "cityName" : "cityName",
            "cityId" : 1
          },
          "locationId" : 6,
          "workingSpots" : [ {
            "workingSpotId" : 5,
            "locationId" : 5,
            "location" : null,
            "workingSpotReservations" : [ {
              "workingSpotId" : 7,
              "workingSpot" : null,
              "workingSpotReservationId" : 2,
              "reservationDate" : "2000-01-23",
              "user" : {
                "jobId" : 2,
                "password" : "",
                "reservations" : [ {
                  "meetingRoom" : null,
                  "meetingRoomReservationId" : 4,
                  "meetingRoomId" : 1,
                  "meetingDate" : "2000-01-23",
                  "startTime" : "startTime",
                  "endTime" : "endTime",
                  "meetingTopic" : "meetingTopic",
                  "user" : null,
                  "userId" : 7
                }, {
                  "meetingRoom" : null,
                  "meetingRoomReservationId" : 4,
                  "meetingRoomId" : 1,
                  "meetingDate" : "2000-01-23",
                  "startTime" : "startTime",
                  "endTime" : "endTime",
                  "meetingTopic" : "meetingTopic",
                  "user" : null,
                  "userId" : 7
                } ],
                "surname" : "surname",
                "name" : "name",
                "job" : {
                  "jobId" : 3,
                  "jobTitle" : "jobTitle"
                },
                "userId" : 9,
                "email" : ""
              },
              "userId" : 1
            }, {
              "workingSpotId" : 7,
              "workingSpot" : null,
              "workingSpotReservationId" : 2,
              "reservationDate" : "2000-01-23",
              "user" : {
                "jobId" : 2,
                "password" : "",
                "reservations" : [ {
                  "meetingRoom" : null,
                  "meetingRoomReservationId" : 4,
                  "meetingRoomId" : 1,
                  "meetingDate" : "2000-01-23",
                  "startTime" : "startTime",
                  "endTime" : "endTime",
                  "meetingTopic" : "meetingTopic",
                  "user" : null,
                  "userId" : 7
                }, {
                  "meetingRoom" : null,
                  "meetingRoomReservationId" : 4,
                  "meetingRoomId" : 1,
                  "meetingDate" : "2000-01-23",
                  "startTime" : "startTime",
                  "endTime" : "endTime",
                  "meetingTopic" : "meetingTopic",
                  "user" : null,
                  "userId" : 7
                } ],
                "surname" : "surname",
                "name" : "name",
                "job" : {
                  "jobId" : 3,
                  "jobTitle" : "jobTitle"
                },
                "userId" : 9,
                "email" : ""
              },
              "userId" : 1
            } ]
          }, {
            "workingSpotId" : 5,
            "locationId" : 5,
            "location" : null,
            "workingSpotReservations" : [ {
              "workingSpotId" : 7,
              "workingSpot" : null,
              "workingSpotReservationId" : 2,
              "reservationDate" : "2000-01-23",
              "user" : {
                "jobId" : 2,
                "password" : "",
                "reservations" : [ {
                  "meetingRoom" : null,
                  "meetingRoomReservationId" : 4,
                  "meetingRoomId" : 1,
                  "meetingDate" : "2000-01-23",
                  "startTime" : "startTime",
                  "endTime" : "endTime",
                  "meetingTopic" : "meetingTopic",
                  "user" : null,
                  "userId" : 7
                }, {
                  "meetingRoom" : null,
                  "meetingRoomReservationId" : 4,
                  "meetingRoomId" : 1,
                  "meetingDate" : "2000-01-23",
                  "startTime" : "startTime",
                  "endTime" : "endTime",
                  "meetingTopic" : "meetingTopic",
                  "user" : null,
                  "userId" : 7
                } ],
                "surname" : "surname",
                "name" : "name",
                "job" : {
                  "jobId" : 3,
                  "jobTitle" : "jobTitle"
                },
                "userId" : 9,
                "email" : ""
              },
              "userId" : 1
            }, {
              "workingSpotId" : 7,
              "workingSpot" : null,
              "workingSpotReservationId" : 2,
              "reservationDate" : "2000-01-23",
              "user" : {
                "jobId" : 2,
                "password" : "",
                "reservations" : [ {
                  "meetingRoom" : null,
                  "meetingRoomReservationId" : 4,
                  "meetingRoomId" : 1,
                  "meetingDate" : "2000-01-23",
                  "startTime" : "startTime",
                  "endTime" : "endTime",
                  "meetingTopic" : "meetingTopic",
                  "user" : null,
                  "userId" : 7
                }, {
                  "meetingRoom" : null,
                  "meetingRoomReservationId" : 4,
                  "meetingRoomId" : 1,
                  "meetingDate" : "2000-01-23",
                  "startTime" : "startTime",
                  "endTime" : "endTime",
                  "meetingTopic" : "meetingTopic",
                  "user" : null,
                  "userId" : 7
                } ],
                "surname" : "surname",
                "name" : "name",
                "job" : {
                  "jobId" : 3,
                  "jobTitle" : "jobTitle"
                },
                "userId" : 9,
                "email" : ""
              },
              "userId" : 1
            } ]
          } ]
        }
      },
...
    }
  }
}